UC_MOVE_REQUEST_COMPUTING_COST finished
